LayoutTests/http/tests/resourceLoadStatistics/remove-blocking-in-redirect-expected.txt

Tests that blocking is removed mid-flight in redirects.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
  

--------
Frame: '<!--frame1-->'
--------
Should receive first-party cookie.
Received cookie named 'firstPartyCookie'.
Did not receive cookie named 'partitionedCookie'.
Client-side document.cookie: firstPartyCookie=value

--------
Frame: '<!--frame2-->'
--------
Redirect case 1, should receive first-party cookie for 127.0.0.1.
Received cookie named 'firstPartyCookie'.
Did not receive cookie named 'partitionedCookie'.
Client-side document.cookie: firstPartyCookie=127.0.0.1

--------
Frame: '<!--frame3-->'
--------
Set partitioned, third-party cookie.


--------
Frame: '<!--frame4-->'
--------
Should receive no cookie.
Did not receive cookie named 'firstPartyCookie'.
Did not receive cookie named 'partitionedCookie'.
Client-side document.cookie:

--------
Frame: '<!--frame5-->'
--------
Redirect case 2, should receive first-party cookie for 127.0.0.1.
Received cookie named 'firstPartyCookie'.
Did not receive cookie named 'partitionedCookie'.
Client-side document.cookie: firstPartyCookie=127.0.0.1
